fix: 퍼센트 소수점으로 수정

+++ src/main/java/com/ssafy/marimo/insurance/dto/response/GetCarInsuranceResponse.java
@@ -26,7 +26,7 @@ public record GetCarInsuranceResponse(
         Float totalDistanceExpectation,
         Integer discountDifferenceWithNextStage,
         Boolean canSubmitDistanceNow,
-        Integer drivingPercentage
+        Float drivingPercentage
 ) {
     public static GetCarInsuranceResponse of(
             String insuranceCompanyName,
@@ -46,7 +46,7 @@ public record GetCarInsuranceResponse(
             Float totalDistanceExpectation,
             Integer discountDifferenceWithNextStage,
             Boolean canSubmitDistanceNow,
-            Integer drivingPercentage
+            Float drivingPercentage
     ) {
         return GetCarInsuranceResponse.builder()
                 .insuranceCompanyName(insuranceCompanyName)
